% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPR_NERR_simple.R
\name{EPR_NERR_simple}
\alias{EPR_NERR_simple}
\title{Internal function to produce SSD and EPR}
\usage{
EPR_NERR_simple(Params)
}
\arguments{
\item{Params}{This function is written such that it only takes one set of M, Linf, and k at a time for a given site.  Params is still a list
with Params$M, Params$Linf, Params$k, but they only contain one value each.  This function also requires Params$veclength and Params$x, 
which are created in run_EPR_NERR_simple.}
}
\description{
This is the internal function that does all the work of producing the EPR and SSD for each site that is fed back to run_EPR_NERR_simple.
}
\examples{
EPR_NERR_simple(Params)
}
\keyword{NERR}
